+++ CHANGELOG.md
@@ -5,12 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 0.4.2 - 2022-03-26
 
 ### Fixed
 
 * Don't fail when modifying/deleting attachment in Confluence Server (#18)
 * User reference now supports `-` symbol in the middle of username, e.g. `~user-name` (#21)
+
+### Changed
+
+* Cleanup of orphaned pages is done in parallel to reduce execution time
 
 ## 0.4.1 - 2022-03-14